requirements.txt
ml-dtypes==0.2.0
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
absl-py==2.1.0
annotated-types==0.7.0
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
backend==0.2.4.1
bitarray==3.0.0
bitstring==4.3.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
clize==5.0.2
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.2
cupy-cuda12x==13.4.1
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
docutils==0.21.2
exceptiongroup==1.3.0
executing==2.2.0
fastrlock==0.8.3
filelock==3.18.0
fixedpoint==1.0.1
flatbuffers==25.2.10
fonttools==4.59.0
fsspec==2025.3.0
gast==0.6.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.71.0
h5py==3.13.0
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
ipykernel==6.29.5
ipython==8.37.0
jedi==0.19.2
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==2.14.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
od==2.0.2
onnx==1.17.0
onnxruntime-gpu==1.21.0
onnxruntime_extensions==0.14.0
onnxscript==0.2.5
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
pip==22.0.2
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==3.20.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyDigitalWaveTools==1.1
pydot==3.0.4
Pygments==2.19.2
pynqmetadata==0.1.5
pynqutils==0.1.1
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
python-magic==0.4.27
pytz==2025.1
PyYAML==6.0.2
pyzmq==27.0.0
qonnx==0.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
setuptools==59.6.0
sigtools==4.0.1
six==1.17.0
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.1
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf2onnx==1.16.1
threadpoolctl==3.6.0
tokenizers==0.21.1
toposort==1.10
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
typing_extensions==4.12.2
typing-inspection==0.4.0
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.14.1
pycuda
